fixed minimap, mouse and added features

+++ include/movement.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   movement.h                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gpasztor <gpasztor@student.42.fr>          +#+  +:+       +#+        */
+/*   By: ffederol <ffederol@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/12 22:38:40 by ffederol          #+#    #+#             */
-/*   Updated: 2023/09/12 18:44:02 by gpasztor         ###   ########.fr       */
+/*   Updated: 2023/09/17 23:53:06 by ffederol         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 # include "player.h"
 # include "cub3d.h"
 
-void	rotate(t_player *player, int id);
+void	rotate(t_player *player, float id);
 void	move_straight(t_player *player, int id, char **worldmap);
 void	move_side(t_player *player, int id, char **worldmap);
 void	interact(mlx_key_data_t keydata, void *param);
